se agrego 2 test e2e  mas para el metodo Get /taxis
